Gift.xcodeproj: Add Gift-iOS target
Fixes #3.

Gift can now be included in iOS projects. The iOS framework is tested in
Gift-iOSTests.

Gift for iOS requires a version of libgit2.a that is built with iOS
architectures in mind. I have no idea how to build such architetures, so
I included the built products directly in this commit. This should be
resolved in issue #9.

Since calls to the system (as well as the NSTask API) is unavailable
from within iOS targets, use soffes/ssziparchive for unzipping fixture
repositories in the test target.